+++ lib/features/paint_editor/paint_editor.dart
@@ -225,9 +225,6 @@ class PaintEditorState extends State<PaintEditor>
   /// A boolean flag representing whether the fill mode is enabled or disabled.
   bool _isFillMode = false;
 
-  /// Controls high-performance for free-style drawing.
-  bool _freeStyleHighPerformance = false;
-
   /// Get the fillBackground status.
   bool get fillBackground => _isFillMode;
 
@@ -876,12 +873,6 @@ class PaintEditorState extends State<PaintEditor>
           zoomConfigs: paintEditorConfigs,
           enableInteraction: paintMode == PaintMode.moveAndZoom,
           onInteractionStart: (details) {
-            _freeStyleHighPerformance =
-                (paintEditorConfigs.enableFreeStyleHighPerformanceMoving ??
-                        !isDesktop) ||
-                    (paintEditorConfigs.enableFreeStyleHighPerformanceScaling ??
-                        !isDesktop);
-
             callbacks.paintEditorCallbacks?.onEditorZoomScaleStart
                 ?.call(details);
             setState(() {});
@@ -889,7 +880,6 @@ class PaintEditorState extends State<PaintEditor>
           onInteractionUpdate:
               callbacks.paintEditorCallbacks?.onEditorZoomScaleUpdate,
           onInteractionEnd: (details) {
-            _freeStyleHighPerformance = false;
             callbacks.paintEditorCallbacks?.onEditorZoomScaleEnd?.call(details);
             setState(() {});
           },
@@ -1011,7 +1001,6 @@ class PaintEditorState extends State<PaintEditor>
       drawAreaSize: mainBodySize ?? editorBodySize,
       editorBodySize: editorBodySize,
       layerStackScaleFactor: _layerStackTransformHelper.scale,
-      freeStyleHighPerformance: _freeStyleHighPerformance,
       layers: activeHistory.layers,
       onTap: (details) =>
           callbacks.paintEditorCallbacks?.onTap?.call(this, details),